+++ src/main/resources/applicationContext.xml
@@ -8,7 +8,7 @@
     <property name="username" value="follower"/>
     <property name="password" value="hello"/>
   </bean>
-  <bean id="userDao" class="UserDao">
+  <bean id="userDao" class="UserDaoJdbc">
     <property name="dataSource" ref="dataSource"/>
   </bean>
   <bean id="jdbcContext" class="JdbcContext">
